+++ module-apps/application-alarm-clock/models/AlarmsModel.cpp
@@ -1,0 +1,76 @@
+// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
+// For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
+
+#include "AlarmsModel.hpp"
+#include "application-alarm-clock/widgets/AlarmItem.hpp"
+#include "application-alarm-clock/widgets/AlarmClockStyle.hpp"
+#include "application-alarm-clock/ApplicationAlarmClock.hpp"
+#include <ListView.hpp>
+
+namespace app::alarmClock
+{
+    AlarmsListItemProvider::AlarmsListItemProvider(Application *app) : DatabaseModel(app)
+    {}
+
+    AlarmsModel::AlarmsModel(app::Application *app, std::shared_ptr<AbstractAlarmsRepository> alarmsRepository)
+        : AlarmsListItemProvider(app), alarmsRepository{std::move(alarmsRepository)}
+    {}
+
+    unsigned int AlarmsModel::requestRecordsCount()
+    {
+        return recordsCount;
+    }
+
+    void AlarmsModel::requestRecords(uint32_t offset, uint32_t limit)
+    {
+        alarmsRepository->getLimited(
+            offset, limit, [this](const std::vector<AlarmsRecord> &records, unsigned int alarmsRepoCount) {
+                return onAlarmsRetrieved(records, alarmsRepoCount);
+            });
+    }
+
+    unsigned int AlarmsModel::getMinimalItemHeight() const
+    {
+        return style::alarmClock::window::item::height;
+    }
+
+    gui::ListItem *AlarmsModel::getItem(gui::Order order)
+    {
+        auto record = getRecord(order);
+        if (record == nullptr) {
+            LOG_DEBUG("Empty record in AlarmsModel::GetItem");
+            return nullptr;
+        }
+
+        auto item               = new gui::AlarmItem(record);
+        item->activatedCallback = [this, record](gui::Item &) {
+            if (record->status == AlarmStatus::Off) {
+                record->status = AlarmStatus::On;
+            }
+            else {
+                record->status = AlarmStatus::Off;
+            }
+            alarmsRepository->update(*record, nullptr);
+            return true;
+        };
+        return item;
+    }
+
+    bool AlarmsModel::updateRecords(std::vector<AlarmsRecord> records)
+    {
+        DatabaseModel::updateRecords(std::move(records));
+        list->onProviderDataUpdate();
+        return true;
+    }
+
+    bool AlarmsModel::onAlarmsRetrieved(const std::vector<AlarmsRecord> &records, unsigned int alarmsRepoCount)
+    {
+        if (recordsCount != alarmsRepoCount) {
+            recordsCount = alarmsRepoCount;
+            list->rebuildList(style::listview::RebuildType::Full, 0, true);
+            return false;
+        }
+
+        return updateRecords(records);
+    }
+} // namespace app::alarmClock